Frontend part of item deletion/hiding system
Add another page showing deleted items only

+++ frontend/js/script.js
@@ -12,7 +12,7 @@ const API_HOST = "http://localhost:3000";
 // stores the latest inventory data, refreshed by loadInventory();
 var INVENTORY_DATA = {};
 
-/* Load the inventory by making an AJAX GET call to the "/inventory" endpoint and populating the HTML */
+/* Load the non-deleted inventory by making an AJAX GET call to the "/inventory" endpoint and populating the HTML */
 function loadInventory() {
   $.ajax(API_HOST + "/inventory", {
     type: "GET",
@@ -29,6 +29,8 @@ function loadInventory() {
     // get all the file options from the server and add them to the list
     for(let itemKey in data) {
       let item = data[itemKey];
+      if(item.deleted !== 0) // only show non-deleted items (where item.deleted == 0)
+        continue;
       INVENTORY_DATA[item.id] = item;
 
       // Append each item to the table, along with the action buttons which have onClick event handlers
@@ -37,6 +39,39 @@ function loadInventory() {
         "</td><td>"+item.quantity.toLocaleString()+"</td>"+"<td>\n" +
         "<button type=\"button\" onClick='editItem(`"+item.id+"`)' class=\"btn btn-success\"><i class=\"fas fa-edit\"></i></button>&nbsp;" +
         "<button type=\"button\" onClick='deleteItem(`"+item.id+"`)' class=\"btn btn-danger\"><i class=\"far fa-trash-alt\"></i></button>" +
+        "</td>"+"</tr>");
+    }
+
+  });
+}
+
+/* Load the deleted inventory by making an AJAX GET call to the "/inventory" endpoint and populating the HTML */
+/* yes, this function could be merged with loadInventory() */
+function loadDeletedInventory() {
+  $.ajax(API_HOST + "/inventory", {
+    type: "GET",
+    dataType: "json",
+    cache: false,
+    json: true,
+  }).done( function (data) {
+    // clear the existing items HTML
+    $("#tableBody").html("");
+
+    // reset the latest data
+    INVENTORY_DATA = {};
+
+    // get all the file options from the server and add them to the list
+    for(let itemKey in data) {
+      let item = data[itemKey];
+      if(item.deleted !== 1) // only show deleted items (where item.deleted == 1)
+        continue;
+      INVENTORY_DATA[item.id] = item;
+
+      // Append each item to the table, along with the action buttons which have onClick event handlers
+      $("#tableBody").append("<tr><th scope='row'>"+item.id+"</th><td>"+
+        item.name+"</td><td>$"+parseFloat(item.cost).toLocaleString('en-US', {minimumFractionDigits: 2, maximumFractionDigits: 2})+
+        "</td><td>"+item.quantity.toLocaleString()+"</td><td>"+item.deletedReason+"</td>"+"<td>\n" +
+        "<button type=\"button\" onClick='restoreItem(`"+item.id+"`)' class=\"btn btn-primary\"><i class=\"fas fa-undo\"></i></button>" +
         "</td>"+"</tr>");
     }
 
@@ -98,19 +133,51 @@ function editItem(itemId) {
   });
 }
 
-/* Delete an item from inventory by making an AJAX DELETE call to the /inventory endpoint with the item ID */
+/* Restore an item by sending a PUT request to the server with identical item details */
+function restoreItem(itemId) {
+  $.ajax(API_HOST + "/inventory", {
+    type: "PUT",
+    data: JSON.stringify({
+      id: itemId,
+      name: INVENTORY_DATA[itemId].name,
+      cost: INVENTORY_DATA[itemId].cost,
+      quantity: INVENTORY_DATA[itemId].quantity,
+    }),
+    contentType: "application/json",
+  }).done( function (data) {
+    if(data.hasOwnProperty("success") && data.success === true)
+      Swal.fire({
+        title: "Item Restored!",
+        text: "Item #"+itemId+" has been successfully un-deleted",
+        icon: "success",
+      });
+
+    // update the list of items
+    loadDeletedInventory();
+  });
+}
+
+/* Delete an item from inventory by making an AJAX DELETE call to the /inventory endpoint with the item ID and a deletion reason */
 function deleteItem(itemId) {
   Swal.fire({
     title: 'Delete Item',
     text: 'Are you sure you want to delete item #'+itemId+'?',
     showCancelButton: true,
     confirmButtonText: 'Delete',
+    html: '<input id="reason" class="swal2-input" placeholder="Reason for Deletion">',
+    focusConfirm: false,
+    preConfirm: () => {
+      return {
+        reason: document.getElementById('reason').value,
+      }
+    },
   }).then((result) => {
-    if(result.isConfirmed) {
+    if(result.isConfirmed && result.value !== undefined && result.value.hasOwnProperty("reason")) {
       $.ajax(API_HOST + "/inventory", {
         type: "DELETE",
         data: JSON.stringify({
           id: itemId,
+          reason: result.value.reason,
         }),
         contentType: "application/json",
       }).done( function (data) {
@@ -152,9 +219,6 @@ function downloadBlob(content, filename, contentType) {
  * RETURNS      :  void
  */
 $(document).ready(function() {
-  // first thing - load the inventory
-  loadInventory();
-
   // new item button functionality
   $("#newItemBtn").click(function() {
     Swal.fire({
